+++ src/Awesomite/StackTrace/VarDumpers/Properties/Properties.php
@@ -7,9 +7,11 @@ use Awesomite\StackTrace\Exceptions\InvalidArgumentException;
 /**
  * @internal
  */
-class Properties implements PropertiesInterface
+class Properties extends AbstractProperties
 {
-    private $object;
+    static private $mapping = array(
+        '\ArrayObject' => '\Awesomite\StackTrace\VarDumpers\Properties\PropertiesArrayObject',
+    );
 
     /**
      * Properties constructor.
@@ -26,49 +28,18 @@ class Properties implements PropertiesInterface
     public function getProperties()
     {
         $object = $this->object;
-        $result = array_map(function ($property) use ($object) {
-            return new ReflectionProperty($property, $object);
-        }, $this->getDeclaredProperties());
 
-        $object = $this->object;
-        if ($object instanceof \ArrayObject) {
-            $result = array_filter($result, function (PropertyInterface $property) use ($object) {
-                return property_exists($object, $property->getName());
-            });
-            $names = array_map(function (PropertyInterface $property) {
-                return $property->getName();
-            }, $result);
-            foreach (get_object_vars($object) as $key => $value) {
-                if (in_array($key, $names)) {
-                    // @codeCoverageIgnoreStart
-                    continue;
-                    // @codeCoverageIgnoreEnd
-                }
-                $result[] = new VarProperty($key, $value);
+        foreach (self::$mapping as $classInterface => $classReader) {
+            if ($object instanceof $classInterface) {
+                /** @var PropertiesInterface $reader */
+                $reader = new $classReader($object);
+
+                return $reader->getProperties();
             }
         }
 
-        return array_values($result);
-    }
-
-    private function getDeclaredProperties()
-    {
-        $reflection = new \ReflectionObject($this->object);
-        $result = array();
-        do {
-            $result += $this->getDeclaredPropertiesForReflection($reflection);
-        } while ($reflection = $reflection->getParentClass());
-
-        return $result;
-    }
-
-    private function getDeclaredPropertiesForReflection(\ReflectionClass $reflection)
-    {
-        $result = array();
-        foreach ($reflection->getProperties() as $property) {
-            $result[$property->getDeclaringClass()->getName() . '__' . $property->getName()] = $property;
-        }
-
-        return $result;
+        return array_map(function ($property) use ($object) {
+            return new ReflectionProperty($property, $object);
+        }, $this->getDeclaredProperties());
     }
 }