+++ src/Awesomite/StackTrace/VarDumpers/Properties/VarProperty.php
@@ -2,29 +2,53 @@
 
 namespace Awesomite\StackTrace\VarDumpers\Properties;
 
-use Awesomite\StackTrace\Exceptions\LogicException;
+use Awesomite\StackTrace\Exceptions\InvalidArgumentException;
 
 /**
  * @internal
  */
 class VarProperty implements PropertyInterface
 {
+    const VISIBILITY_DYNAMIC = 'dynamic';
+    const VISIBILITY_PUBLIC = 'public';
+    const VISIBILITY_PROTECTED = 'protected';
+    const VISIBILITY_PRIVATE = 'private';
+
     private $name;
+
+    private $value;
+
+    private $static;
+
+    private $visibility;
+
+    private $declaringClass;
 
     /**
      * VarProperty constructor.
      * @param string $name
      * @param $value
+     * @param string $visibility
+     * @param string $declaringClass
+     * @param bool $static
      */
-    public function __construct($name, $value)
+    public function __construct($name, $value, $visibility, $declaringClass, $static = false)
     {
+        $visibilityOptions = array(
+            static::VISIBILITY_DYNAMIC,
+            static::VISIBILITY_PUBLIC,
+            static::VISIBILITY_PROTECTED,
+            static::VISIBILITY_PRIVATE,
+        );
+        if (!in_array($visibility, $visibilityOptions, true)) {
+            throw new InvalidArgumentException('Invalid value for $visibility!');
+        }
+
         $this->name = $name;
         $this->value = $value;
-    }
-
-    public function hasReflection()
-    {
-        return false;
+        $this->visibility = $visibility;
+        $this->declaringClass = $declaringClass;
+        $this->static = $static;
     }
 
     public function getName()
@@ -32,13 +56,38 @@ class VarProperty implements PropertyInterface
         return $this->name;
     }
 
-    public function getReflection()
-    {
-        throw new LogicException("Reflection for property {$this->name} does not exist!");
-    }
-
     public function getValue()
     {
         return $this->value;
+    }
+
+    public function isStatic()
+    {
+        return $this->static;
+    }
+
+    public function isDynamic()
+    {
+        return $this->visibility === static::VISIBILITY_DYNAMIC;
+    }
+
+    public function isPublic()
+    {
+        return $this->visibility === static::VISIBILITY_PUBLIC;
+    }
+
+    public function isProtected()
+    {
+        return $this->visibility === static::VISIBILITY_PROTECTED;
+    }
+
+    public function isPrivate()
+    {
+        return $this->visibility === static::VISIBILITY_PRIVATE;
+    }
+
+    public function getDeclaringClass()
+    {
+        return $this->declaringClass;
     }
 }